Moving styles outside the component.
