+++ src/components/Login/Login.styled.ts
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+import TextField from '@material-ui/core/TextField';
+
+const TextFieldSize = styled(TextField)`
+  width: 332px;
+`;
+
+const Container = styled.div`
+  display: grid;
+`;
+
+const Styled = {
+  TextFieldSize,
+  Container,
+};
+
+export default Styled;
