+++ src/components/Mosaic/Mosaic.styled.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const CardMediaRoot = styled.div`
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(365px, 1fr));
+  margin: 32px;
+  justify-items: center;
+`;
+
+const Styled = {
+  CardMediaRoot
+};
+
+export default Styled;
