+++ src/components/NotFound/NotFound.styled.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Container = styled.div`
+  margin: 32px;
+`;
+
+const Styled = {
+  Container,
+};
+
+export default Styled;
